refactor!: put crypto layer additions into separate folder and rewrote `getProvider` function
BREAKING CHANGE: `getProvider` takes `ProviderIdentifier` type.
