+++ src/index.ts
@@ -1,12 +1,12 @@
 export * from "./BaseX";
 export * from "./buildInformation";
 export * from "./CoreBuffer";
+export * from "./crypto-layer/CryptoLayerConfig";
+export * from "./crypto-layer/CryptoPrivateKeyHandle";
 export * from "./CryptoDerivation";
 export * from "./CryptoError";
 export * from "./CryptoErrorCode";
-export * from "./CryptoLayerConfig";
 export * from "./CryptoPrivateKey";
-export * from "./CryptoPrivateKeyHandle";
 export * from "./CryptoPublicKey";
 export * from "./CryptoValidation";
 export * from "./encryption/CryptoCipher";
